@@ -1,23 +1,23 @@
 name: flutter_clean_architecture
 description: A Flutter package that implements the Clean Architecture by Uncle Bob in Flutter. It provides Views, Controllers, Presenters, Observers, and UseCases.
-version: 4.1.1
+version: 5.0.0
 author: Shady Boukhary <sb199898.sb@gmail.com>
 homepage: https://github.com/ShadyBoukhary/flutter_clean_architecture
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  logging: ^0.11.4
-  meta: ^1.3.0-nullsafety.3
-  provider: ^4.3.2+3
-  responsive_builder: ^0.3.0
-  rxdart: ^0.25.0
+  logging: ^1.0.0
+  meta: ^1.3.0
+  provider: ^5.0.0
+  responsive_builder: ^0.4.0-nullsafety.1
+  rxdart: ^0.26.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  test: ^1.16.0-nullsafety.5
-  pedantic: ^1.10.0-nullsafety.2
+  test: ^1.16.5
+  pedantic: ^1.11.0
