@@ -66,7 +66,7 @@
 ///       }
 /// ```
 abstract class Observer<T> {
-  void onNext(T response);
+  void onNext(T? response);
   void onComplete();
   void onError(e);
 }
